create new folder for models with individual files for each type
