+++ samples/CosmosDBIntegration/ReadThroughSamples/Models/CosmosDBListData.cs
@@ -1,0 +1,9 @@
+﻿ using System.Collections.Generic;
+
+namespace Microsoft.Azure.WebJobs.Extensions.Redis.Samples.CosmosDB.Models
+{
+    public record CosmosDBListData(
+        string id,
+        List<string> value
+        );
+}
