+++ samples/CosmosDBIntegration/ReadThroughSamples/Models/RedisData.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Microsoft.Azure.WebJobs.Extensions.Redis.Samples.CosmosDB.Models
+{
+    public record RedisData(
+        string id,
+        string key,
+        string value,
+        DateTime timestamp
+        );
+}
