+++ samples/CosmosDBIntegration/ReadThroughSamples/Models/StreamDataSingleDocument.cs
@@ -1,29 +1,10 @@
 ﻿using Microsoft.Extensions.Logging;
 using StackExchange.Redis;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Microsoft.Azure.WebJobs.Extensions.Redis.Samples.CosmosDB
+namespace Microsoft.Azure.WebJobs.Extensions.Redis.Samples.CosmosDB.Models
 {
-    public class StreamData
-    {
-        public string id { get; set; }
-        public Dictionary<string, string> values { get; set; }
-
-        // Helper method to format stream message
-        public static StreamData Format(StreamEntry entry, ILogger logger)
-        {
-            logger.LogInformation("ID: {val}", entry.Id.ToString());
-
-            // Map each key value pair
-            Dictionary<string, string> dict = entry.Values.ToDictionary(value => value.Name.ToString(), value => value.Value.ToString());
-
-            StreamData sampleItem = new StreamData { id = entry.Id, values = dict };
-            return sampleItem;
-        }
-    }
-
     public class StreamDataSingleDocument
     {
         public string id { get; set; }
@@ -65,25 +46,6 @@ namespace Microsoft.Azure.WebJobs.Extensions.Redis.Samples.CosmosDB
             StreamDataSingleDocument data = new StreamDataSingleDocument { id = results.id, maxlen = results.maxlen, messages = list };
             return data;
         }
+
     }
-
-    public record RedisData(
-        string id,
-        string key,
-        string value,
-        DateTime timestamp
-        );
-
-    public record PubSubData(
-        string id,
-        string channel,
-        string message,
-        DateTime timestamp
-        );
-
-    public record CosmosDBListData(
-        string id,
-        List<string> value
-        );
-
 }
