+++ samples/CosmosDBIntegration/ReadThroughSamples/ReadThrough/ListReadThrough.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.Cosmos.Linq;
+using Microsoft.Azure.WebJobs.Extensions.Redis.Samples.CosmosDB.Models;
 using Microsoft.Extensions.Logging;
 using StackExchange.Redis;
 using System;
