+++ samples/CosmosDBIntegration/ReadThroughSamples/WriteAround/PubSubWriteAround.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using Microsoft.Azure.WebJobs.Extensions.Redis.Samples.CosmosDB.Models;
+using Microsoft.Extensions.Logging;
 using StackExchange.Redis;
 using System;
 using System.Collections.Generic;
